Fixes execution order and some commands
- Setting appropriate and secure permissions and ownership for
files/folders is only executed new instance/theme/plugin is
installed/updated, and some internal directories have been excluded from
this process.
- Donwloaded tar is now kept in the installation dir,
except when no version provided (this is so because there is no way of
guessing which version of WordPress is installed).
- Harder security
permissions for `wp-config.php` (0640).
- Execution order and
dependencies between resources fixed and optimized.
- Code refactored
and optimized.